MUS-152 Make use of Gradientdrawable, Refactor test
This commit deletes the png-based images which were used to draw the white piano keyboard and replaces them with GradientDrawables as these should scale better for different form factors. Also the geometry of the buttons was adjusted a bit to visually match better with a piano keyboard. Therefore also the UI unit test for button positioning had to be adjusted a bit.